Done coursework 7 (#10)
* Modify workflows/action.yaml

* Add 'pytest' package

* Implement UserLoginView tests

* Implement UserRetrieveView tests

* Implement UserCreateView tests

* Implement UpdatePasswordView tests

* Implement tests for POST: board/create

* Implement tests for GET, PATCH, DELETE: board/<pk>

* Implement tests for GET: board/list

* Implement tests for GET: category-list

* Implement tests for POST: category-create

* Implement tests for [GET, PATCH, DELETE]: category-detail

* Implement tests for GET: goal-list

* Implement tests for POST: goal-create

* Implement tests for [GET, PATCH, DELETE]: goal-detail

* Implement tests for GET: comment-list

* Implement tests for POST: comment-create

* Implement tests for [GET, PATCH, DELETE]: comment-detail

* Implement tests for GET: health-check

* Modify docstrings

* Add docstrings

* Modify README.md

* Add .env.example file

* Create multi-stage build Dockerfile

* Modify README.md

+++ todolist/goals/serializers.py
@@ -1,5 +1,5 @@
 from django.db import transaction
-from rest_framework import serializers
+from rest_framework import serializers, exceptions
 
 from core.models import User
 from core.serializers import ProfileSerializer
@@ -7,8 +7,13 @@ from goals.models import Category, Goal, Comment, Board, BoardParticipant
 
 
 class BoardCreateSerializer(serializers.ModelSerializer):
+    """Сериализатор представления BoardViewSet
+
+    Action create
+    """
     user = serializers.HiddenField(default='user')
 
+    #: Добавление текущего пользователя владельцем в список участников доски
     def create(self, validated_data) -> Board:
         user = validated_data.pop('user')
         board = Board.objects.create(**validated_data)
@@ -22,6 +27,10 @@ class BoardCreateSerializer(serializers.ModelSerializer):
 
 
 class BoardParticipantSerializer(serializers.ModelSerializer):
+    """Сериализатор модели BoardParticipant
+
+    Для преобразования данных об участниках доски
+    """
     role = serializers.ChoiceField(required=True, choices=BoardParticipant.Role.choices[1:])
     user = serializers.SlugRelatedField(slug_field='username', queryset=User.objects.all())
 
@@ -32,6 +41,10 @@ class BoardParticipantSerializer(serializers.ModelSerializer):
 
 
 class BoardUpdateSerializer(serializers.ModelSerializer):
+    """Сериализатор представления BoardViewSet
+
+    Actions: update, retrieve, partial_update, destroy
+    """
     participants = BoardParticipantSerializer(many=True)
     user = serializers.HiddenField(default='user')
 
@@ -40,52 +53,60 @@ class BoardUpdateSerializer(serializers.ModelSerializer):
         fields = '__all__'
         read_only_fields = ('id', 'created', 'updated',)
 
+    #: Реализация частичного обновления доски
     def update(self, instance: Board, validated_data: dict) -> Board:
         owner = validated_data.pop('user')
-        new_participants = validated_data.pop('participants')
-        new_by_id = {part['user'].id: part for part in new_participants}
-        old_participants = instance.participants.exclude(user=owner)
-        with transaction.atomic():
-            for old_participant in old_participants:
-                if old_participant.user_id not in new_by_id:
-                    old_participant.delete()
-                else:
-                    if old_participant.role != new_by_id[old_participant.user_id]['role']:
-                        old_participant.role = new_by_id[old_participant.user_id]['role']
-                        old_participant.save()
-                    new_by_id.pop(old_participant.user_id)
-            for new_part in new_by_id.values():
-                BoardParticipant.objects.create(
-                    user=new_part['user'], board=instance, role=new_part['role']
-                )
-            if title := validated_data.get('title'):
-                instance.title = title
+        if new_participants := validated_data.get('participants'):
+            new_by_id = {part['user'].id: part for part in new_participants}
+            old_participants = instance.participants.exclude(user=owner)
+            with transaction.atomic():
+                for old_participant in old_participants:
+                    if old_participant.user_id not in new_by_id:
+                        old_participant.delete()
+                    else:
+                        if old_participant.role != new_by_id[old_participant.user_id]['role']:
+                            old_participant.role = new_by_id[old_participant.user_id]['role']
+                            old_participant.save()
+                        new_by_id.pop(old_participant.user_id)
+                for new_part in new_by_id.values():
+                    BoardParticipant.objects.create(
+                        user=new_part['user'], board=instance, role=new_part['role']
+                    )
+        if title := validated_data.get('title'):
+            instance.title = title
 
-            instance.save()
+        instance.save()
 
         return instance
 
 
 class BoardListSerializer(serializers.ModelSerializer):
+    """Сериализатор представления BoardViewSet
 
+    Action list
+    """
     class Meta:
         model = Board
         fields = '__all__'
 
 
 class CategoryCreateSerializer(serializers.ModelSerializer):
+    """Сериализатор представления CategoryViewSet
+
+    Action create
+    """
     user = serializers.HiddenField(default='user')
 
     def validate_board(self, value: Board) -> Board:
-        # Проверка статуса доски 'Удалена'
+        #: Проверка статуса доски
         if value.is_deleted:
             raise serializers.ValidationError('Not allowed in deleted category')
-        # Проверка роли пользователя
+        #: Проверка роли пользователя
         if not value.participants.filter(
             user_id=self.context['request'].user.id,
             role__in=(BoardParticipant.Role.owner, BoardParticipant.Role.writer,)
         ).exists():
-            raise serializers.ValidationError('Not owner or writer of category')
+            raise exceptions.PermissionDenied
 
         return value
 
@@ -96,6 +117,10 @@ class CategoryCreateSerializer(serializers.ModelSerializer):
 
 
 class CategoryListSerializer(serializers.ModelSerializer):
+    """Сериализатор представления CategoryViewSet
+
+    Actions: list, update, retrieve, partial_update, destroy
+    """
     user = ProfileSerializer(read_only=True)
 
     class Meta:
@@ -105,18 +130,22 @@ class CategoryListSerializer(serializers.ModelSerializer):
 
 
 class GoalCreateSerializer(serializers.ModelSerializer):
+    """Сериализатор представления GoalViewSet
+
+    Action create
+    """
     user = serializers.HiddenField(default='user')
 
     def validate_category(self, value: Category) -> Category:
-        # Проверка статуса категории 'Удалена'
+        #: Проверка статуса категории
         if value.is_deleted:
             raise serializers.ValidationError('Not allowed in deleted category')
-        # Проверка роли пользователя
+        #: Проверка роли пользователя
         if not value.board.participants.filter(
                 user_id=self.context['request'].user.id,
                 role__in=(BoardParticipant.Role.owner, BoardParticipant.Role.writer,)
         ).exists():
-            raise serializers.ValidationError('Not owner or writer of category')
+            raise exceptions.PermissionDenied
 
         return value
 
@@ -127,6 +156,10 @@ class GoalCreateSerializer(serializers.ModelSerializer):
 
 
 class GoalListSerializer(serializers.ModelSerializer):
+    """Сериализатор представления GoalViewSet
+
+    Actions: list, update, retrieve, partial_update, destroy
+    """
     user = ProfileSerializer(read_only=True)
 
     class Meta:
@@ -136,18 +169,22 @@ class GoalListSerializer(serializers.ModelSerializer):
 
 
 class CommentCreateSerializer(serializers.ModelSerializer):
+    """Сериализатор представления CommentViewSet
+
+    Action create
+    """
     user = serializers.HiddenField(default='user')
 
     def validate_goal(self, value: Goal) -> Goal:
-        # Проверка статуса цели 'Архив'
+        #: Проверка статуса цели
         if value.status == Goal.Status.archived:
             raise serializers.ValidationError('Not allowed in archived goal')
-        # Проверка роли пользователя
+        #: Проверка роли пользователя
         if not value.category.board.participants.filter(
                 user_id=self.context['request'].user.id,
                 role__in=(BoardParticipant.Role.owner, BoardParticipant.Role.writer,)
         ).exists():
-            raise serializers.ValidationError('Not owner or writer of goal')
+            raise exceptions.PermissionDenied
 
         return value
 
@@ -158,6 +195,10 @@ class CommentCreateSerializer(serializers.ModelSerializer):
 
 
 class CommentListSerializer(serializers.ModelSerializer):
+    """Сериализатор представления CommentViewSet
+
+    Actions: list, update, retrieve, partial_update, destroy
+    """
     user = ProfileSerializer(read_only=True)
 
     class Meta:
